Ajout bouton confirmation

+++ admin/templates/articles/blog.html.php
@@ -25,7 +25,7 @@ foreach($articles as $article) : { ?>
                             <div class="card-header"><small class="text-muted">Dernière mise à jour <?= $article['created_at'] ?></small></div>
                             <p class="card-text"><?= $article['chapo'] ?></p>
                             <a class="btn btn-info" href="index.php?controller=article&task=show&id=<?= $article['id'] ?>">Modifier cet article</a>
-                            <a class="btn btn-info" href="index.php?controller=article&task=delete&id=<?= $article['id'] ?>">Supprimer cet article</a>
+                            <a class="btn btn-info" onclick="return confirm('Êtes vous sur de vouloir supprimer l\'article ?')" href="index.php?controller=article&task=delete&id=<?= $article['id'] ?>">Supprimer cet article</a>
                         </div>
                     </div>
             </div>
